Completed browser upload and some QOL. TODO: more testing
