+++ hp_tool/data/db.csv
@@ -1,129 +1,129 @@
-DeviceType,Manufacturer,LensMount
-360 cam,Adobe Systems,builtin 
-Action Camera,AIPTEK,Canon EF
-Analog,Apple,Canon EF-M
-AudioRecorder,Argus,Canon EF-S
-Autonomous/Other Car Camera,AUBBC,Canon EX
-Body Camera,Audiovox,Canon FD
-Camcorder,Axis,Canon FL
-CellPhone,Bell & Howell,Fujifilm X-Mount
-Compact Digital,Betterlight,Hasselblad V-Mount
-Dashcam,BlackBerry,Leica Camera L-Mount
-Digital Back,Blackmagic Design,Leica Camera M Mount
-Drone Camera,BLU,Leica Camera R Bayonet
-DSLR,Bosch,Leica Camera S-Mount
-Film Scanner,Bowens,Mamiya Bayonet
-Flatbed Scanner,Broncolor,Nikon 1-Mount
-InkJetPrinter,Cambo,Nikon AF-S
-LaserPrinter,CamKix,Nikon F-Mount
-Lens,Canfelid,Nikon S-Mount
-Light Field,Canon,Olympus Four Thirds
-Medium Format,Casio,Olympus OM-Mount
-Mirrorless,Chimera,Pentax K
-Multi Lens Mobile Phone,Cobra,Pentax Q
-Multi sensor camera,Concord Camera Corp,Samsung NX Mount
-Multi Spectral,Creative,Sigma SA
-MultiSpectral,Critical Link LLC,Sony E Mount
-Pen Camera,DearDroff Cameras,Sony FZ-mount
-PhotographicPrinter,default,
-Point&Shoot,Digital Concepts,
-Raspberry Pi Cam,DJI,
-Remote,DXG Technology,
-Scientific,Edgertronic,
-Security,Efun,
-Stabilization,Epson,
-Tablet,Ezonics,
-Thermal,FREDI,
-ThermalPrinter,FujiFilm,
-Trail Cam,Gadget Universe,
-Underwater,Garmin,
-Unknown,Gateway,
-Virtual Reality,Google,
-Wearable Camera,GoPro,
-Webcam,Gosky,
-,GrandVision,
-,Hamamatsu Corporation,
-,Hasselblad,
-,Hawking Technology,
-,Hikvision,
-,Hisense,
-,HORIBA Scientific,
-,HP (Hewlett -Packard),
-,HTC,
-,Huawei,
-,iClick,
-,iSonic,
-,Jazz,
-,JVC,
-,KBGear,
-,Kino Flo,
-,Kobian,
-,Kodak,
-,Konaki,
-,Konica Minolta,
-,Kyocera,
-,Labtec,
-,Largan,
-,Leica Camera,
-,LG,
-,Logitech,
-,Lowel,
-,Lytro,
-,Mamiya,
-,Matsonic,
-,Mattel,
-,Micro Innovations,
-,Minox,
-,Motorola,
-,Nikon,
-,NISIS,
-,Noblex,
-,NOKIA,
-,Norcent,
-,Olympus,
-,OnePlus,
-,Oregon Scientific,
-,Panasonic,
-,Pelco,
-,Pentax,
-,Phase One,
-,Philips,
-,Photoflex,
-,Photogenic,
-,Pocket Wizard,
-,Polaroid,
-,Profoto,
-,Quaketronics,
-,Quantum,
-,RCA,
-,Rexing,
-,Ricoh,
-,Rollei,
-,Sakar,
-,Samsung,
-,Sanyo,
-,SeaLife,
-,Seek Thermal,
-,Sharp,
-,Sigma,
-,Sinar,
-,SiPix,
-,Sony,
-,Spectra Merchandising International,
-,Thinkware,
-,Toshiba,
-,Toyo,
-,UMAX,
-,uniden,
-,Uokoo,
-,Veilux,
-,Veo,
-,Vie Vu,
-,Vision Research,
-,Vivitar,
-,Westcott,
-,Wild Game,
-,Wolverine Data,
-,Zeiss,
-,Zoom,
-,ZOOM Corporation,
+DeviceType,Manufacturer,LensMount,,,,
+360 Camera,Adobe Systems,builtin ,,,,
+Action Camera,AIPTEK,Canon EF,,,,
+Analog,Apple,Canon EF-M,,,,
+AudioRecorder,Argus,Canon EF-S,,,,
+Autonomous/Other Car Camera,AUBBC,Canon EX,,,,
+Body Camera,Audiovox,Canon FD,,,,
+Camcorder,Axis,Canon FL,,,,
+CellPhone,Bell & Howell,Fujifilm X-Mount,,,,
+Compact Digital,Betterlight,Hasselblad V-Mount,,,,
+Dashcam,BlackBerry,Leica Camera L-Mount,,,,
+Digital Back,Blackmagic Design,Leica Camera M Mount,,,,
+Drone Camera,BLU,Leica Camera R Bayonet,,,,
+DSLR,Bosch,Leica Camera S-Mount,,,,
+Film Scanner,Bowens,Mamiya Bayonet,,,,
+Flatbed Scanner,Broncolor,Nikon 1-Mount,,,,
+InkJetPrinter,Cambo,Nikon AF-S,,,,
+LaserPrinter,CamKix,Nikon F-Mount,,,,
+Lens,Canfelid,Nikon S-Mount,,,,
+Light Field,Canon,Olympus Four Thirds,,,,
+Medium Format,Casio,Olympus OM-Mount,,,,
+Mirrorless,Chimera,Pentax K,,,,
+Multi-Lens Mobile Phone,Cobra,Pentax Q,,,,
+Multi-sensor camera,Concord Camera Corp,Samsung NX Mount,,,,
+MultiSpectral,Creative,Sigma SA,,,,
+Other,Critical Link LLC,Sony E Mount,,,, 
+Pen Camera,DearDroff Cameras,Sony FZ-mount,,,,
+Photographic Printer,default,,,,,
+Point & Shoot,Digital Concepts,,,,,
+Raspberry Pi Camera,DJI,,,,,
+Remote,DXG Technology,,,,,
+Scientific,Edgertronic,,,,,
+Security,Efun,,,,,
+Stabilization,Epson,,,,,
+Tablet,Ezonics,,,,,
+Thermal,FREDI,,,,,
+Thermal Printer,FujiFilm,,,,,
+Trail Cam,Gadget Universe,,,,,
+Underwater,Garmin,,,,,
+Virtual Reality,Gateway,,,,,
+Wearable Camera,Google,,,,,
+Webcam,GoPro,,,,,
+,Gosky,,,,,
+,GrandVision,,,,,
+,Hamamatsu Corporation,,,,,
+,Hasselblad,,,,,
+,Hawking Technology,,,,,
+,Hikvision,,,,,
+,Hisense,,,,,
+,HORIBA Scientific,,,,,
+,HP (Hewlett -Packard),,,,,
+,HTC,,,,,
+,Huawei,,,,,
+,iClick,,,,,
+,iSonic,,,,,
+,Jazz,,,,,
+,JVC,,,,,
+,KBGear,,,,,
+,Kino Flo,,,,,
+,Kobian,,,,,
+,Kodak,,,,,
+,Konaki,,,,,
+,Konica Minolta,,,,,
+,Kyocera,,,,,
+,Labtec,,,,,
+,Largan,,,,,
+,Leica Camera,,,,,
+,LG,,,,,
+,Logitech,,,,,
+,Lowel,,,,,
+,Lytro,,,,,
+,Mamiya,,,,,
+,Matsonic,,,,,
+,Mattel,,,,,
+,Micro Innovations,,,,,
+,Minox,,,,,
+,Motorola,,,,,
+,Nikon,,,,,
+,NISIS,,,,,
+,Noblex,,,,,
+,NOKIA,,,,,
+,Norcent,,,,,
+,Olympus,,,,,
+,OnePlus,,,,,
+,Oregon Scientific,,,,,
+,Panasonic,,,,,
+,Pelco,,,,,
+,Pentax,,,,,
+,Phase One,,,,,
+,Philips,,,,,
+,Photoflex,,,,,
+,Photogenic,,,,,
+,Pocket Wizard,,,,,
+,Polaroid,,,,,
+,Profoto,,,,,
+,Quaketronics,,,,,
+,Quantum,,,,,
+,RCA,,,,,
+,Rexing,,,,,
+,Ricoh,,,,,
+,Rollei,,,,,
+,Sakar,,,,,
+,Samsung,,,,,
+,Sanyo,,,,,
+,SeaLife,,,,,
+,Seek Thermal,,,,,
+,Sharp,,,,,
+,Sigma,,,,,
+,Sinar,,,,,
+,SiPix,,,,,
+,Sony,,,,,
+,Spectra Merchandising International,,,,,
+,Thinkware,,,,,
+,Toshiba,,,,,
+,Toyo,,,,,
+,UMAX,,,,,
+,uniden,,,,,
+,Uokoo,,,,,
+,Veilux,,,,,
+,Veo,,,,,
+,Vie Vu,,,,,
+,Vision Research,,,,,
+,Vivitar,,,,,
+,Westcott,,,,,
+,Wild Game,,,,,
+,Wolverine Data,,,,,
+,Zeiss,,,,,
+,Zoom,,,,,
+,ZOOM Corporation,,,,,
